Updated re2 to the latest version [#CLICKHOUSE-2]

+++ dbms/src/Functions/FunctionsStringSearch.cpp
@@ -623,7 +623,7 @@ struct ReplaceRegexpImpl
     {
         re2_st::StringPiece matches[max_captures];
 
-        int start_pos = 0;
+        size_t start_pos = 0;
         while (start_pos < input.length())
         {
             /// If no more replacements possible for current string
